LibRaw

Documentation for LibRaw.

LibRaw.camera_multiplier

LibRaw.color_description

LibRaw.color_index

LibRaw.demoisaic

LibRaw.height

LibRaw.iparams

LibRaw.raw_height

LibRaw.raw_image

LibRaw.raw_width

LibRaw.subtract_black!

LibRaw.unpack!

LibRaw.width

LibRaw.camera_multiplier — Method

camera_multiplier(image:RawImage)

Return of vector of length 4 containing the camera multipliers https://www.libraw.org/node/2411

LibRaw.color_description — Method

color_description(image::RawImage)

Description of colors numbered from 0 to 3 (RGBG,RGBE,GMCY, or GBTG).

LibRaw.color_index — Method

LibRaw.color_index(image::LibRawImage)

Returns a matrix with the same dimension as the RAW image containing the color index.

LibRaw.demoisaic — Method

demoisaic(algorithm::T, image::RawImage) where T <: DemosaicMethod

Demoisaic the raw imageusingalgorithmmethod (e.g.BayerAverage`).

The resulting array will contains 4 channels (e.g. it will have two green channels), use LibRaw.color_description to get the description of the channels.

LibRaw.height — Method

LibRaw.height(image::RawImage)

returns height of the image (height).

LibRaw.iparams — Method

iparams(image::RawImage)

Returns iparams struct from libraw_data_t.

LibRaw.raw_height — Method

LibRaw.raw_height(image::RawImage)

returns raw height of the image.

LibRaw.raw_image — Method

LibRaw.raw_image(image::RawImage)

Wraps the raw_image field of libraw_rawdata_t in an Array. The data might become corrupted when image is GC'ed.

LibRaw.raw_width — Method

LibRaw.raw_width(image::RawImage)

returns raw width of image.

LibRaw.subtract_black! — Method

LibRaw.subtract_black!(image::RawImage)

This call will subtract black level values from RAW data (for suitable RAW data).

LibRaw.unpack! — Method

LibRaw.unpack!(image::RawImage)

Unpacks the RAW files of the image, calculates the black level.

LibRaw.width — Method

LibRaw.width(image::RawImage)

returns width of the image (iwidth).