LibRaw

Documentation for LibRaw.

LibRaw.RawImage

Base.maximum

LibRaw.black_level

LibRaw.black_level_per_channel

LibRaw.camera_list

LibRaw.camera_multipliers

LibRaw.camera_rgb

LibRaw.color_description

LibRaw.color_index

LibRaw.colordata

LibRaw.demoisaic

LibRaw.height

LibRaw.iparams

LibRaw.pre_multipliers

LibRaw.raw_height

LibRaw.raw_image

LibRaw.raw_width

LibRaw.subtract_black!

LibRaw.unpack!

LibRaw.width

LibRaw.RawImage — Method

RawImage(path::String; unpack = true)

Constructor for the RawImage struct, taking a path to a raw image file as input.

Base.maximum — Method

maximum(image::RawImage)

Maximum pixel value. Calculated from the data for most cameras, hardcoded for others. This value may be changed on postprocessing stage (when black subtraction performed) and by automated maximum adjustment.

LibRaw.black_level — Method

black_level(image::RawImage)

Black level. Depending on the camera, it may be zero (this means that black has been subtracted at the unpacking stage or by the camera itself), calculated at the unpacking stage, read from the RAW file, or hardcoded.

LibRaw.black_level_per_channel — Method

black_level_per_channel(image::RawImage)

! This returns the 4 first values in the `cblack` field of `libraw_colordata_t`. 
Libraw seems to do extra processing to populate these values, see :
https://github.com/letmaik/rawpy/blob/e2f171a0a50053209938e37de3ae0b5fcf6daf08/rawpy/data_helper.h#L11

LibRaw.camera_list — Method

camera_list()

Returns list of supported cameras.

LibRaw.camera_multipliers — Method

camera_multipliers(image:RawImage)

White balance coefficients (as shot). Either read from file or calculated.

Returns of vector of length 4 containing the camera multipliers. https://www.libraw.org/node/2411

LibRaw.camera_rgb — Method

camera_rgb(image:RawImage)

Camera to sRGB conversion matrix.

Returns of a 3x3 matrix (last row is dropped).

LibRaw.color_description — Method

color_description(image::RawImage)

Description of colors numbered from 0 to 3 (RGBG,RGBE,GMCY, or GBTG).

LibRaw.color_index — Method

LibRaw.color_index(image::LibRawImage)

Returns a matrix with the same dimension as the RAW image containing the color index.

LibRaw.colordata — Method

colordata(image::RawImage)

Returns a pointer to colordata field of libraw_data_t.

LibRaw.demoisaic — Method

demoisaic(algorithm::T, image::RawImage) where T <: DemosaicMethod

Demoisaic the raw image using algorithm method (e.g. BayerAverage).

The resulting array will contains 4 channels (e.g. it will have two green channels), use LibRaw.color_description to get the description of the channels.

LibRaw.height — Method

LibRaw.height(image::RawImage)

returns height of the image (height).

LibRaw.iparams — Method

iparams(image::RawImage)

Returns iparams struct from libraw_data_t.

LibRaw.pre_multipliers — Method

pre_multipliers(image:RawImage)

White balance coefficients for daylight (daylight balance). Either read from file, or calculated on the basis of file data, or taken from hardcoded constants.

Returns a vector of length 4 containing the pre-multipliers.

LibRaw.raw_height — Method

LibRaw.raw_height(image::RawImage)

returns raw height of the image.

LibRaw.raw_image — Method

LibRaw.raw_image(image::RawImage)

Wraps the raw_image field of libraw_rawdata_t in an Array. The data might become corrupted when image is GC'ed.

LibRaw.raw_width — Method

LibRaw.raw_width(image::RawImage)

returns raw width of image.

LibRaw.subtract_black! — Method

LibRaw.subtract_black!(image::RawImage)

This call will subtract black level values from RAW data (for suitable RAW data).

LibRaw.unpack! — Method

LibRaw.unpack!(image::RawImage)

Unpacks the RAW files of the image, calculates the black level.

LibRaw.width — Method

LibRaw.width(image::RawImage)

returns width of the image (iwidth).